+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Improved fetch from RSS via CURL 
 - Improved fetch from RSS progress notifications
 - Improved RSS generation for podcasts
+- Make internal podcast handle unique which is now combination of {podcastId}-{podcastSlug} instead of {podcastSlug}
 - Fixed route for episodes
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 - Fixed permission issues where non-admin users were not able to access some plugin actions